Set types for non-metis builds
In case we are not building with the metis library, we still need to set
`metis_*_kind` since those types are used even without the metis
library. It would be helpful to clean that up a bit better and
completely separate the metis type issue from the rest of the code that
doesn't use metis.

Fixes: #43