Fix OpenMPI start options
More recent OpenMPI version start MPI processes in a tree which means
that ssh key login needs to be enabled in between nodes. The added
option turns this feature off until we figure out a better way to
handling this.

Maybe we can tune this by adding some option to ssh to skip host key
verification.